video-processing/requirements.txt
accelerate==1.9.0
boto3==1.39.12
PyYAML==6.0.2 
botocore==1.39.12
certifi==2025.7.14
charset-normalizer==3.4.2
filelock==3.13.1
fsspec==2024.6.1
hf-xet==1.1.5
huggingface-hub==0.33.5
idna==3.10
Jinja2==3.1.4
jmespath==1.0.1
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==2.1.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.1.105
nvidia-nvtx-cu12==12.1.105
opencv-python-headless==4.12.0.88
packaging==25.0
pandas==2.3.1
pillow==11.0.0
psutil==7.0.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
s3transfer==0.13.1
safetensors==0.5.3
sentencepiece==0.2.0
six==1.17.0
sympy==1.13.1
tokenizers==0.21.2
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.1
transformers==4.53.3
triton==3.1.0
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.5.0
